% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesman.R
\name{distance}
\alias{distance}
\title{Euclidean distance between two 2d points}
\usage{
distance(v1, v2)
}
\arguments{
\item{v1}{first (X1,Y1) point}

\item{v2}{second (X1,Y1) point}
}
\value{
a number = sqrt((X1-X2)^2 + (Y1-Y2)^2)
}
\description{
Computing L2 distance between 2 points
}
