% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesman.R
\name{monVoyageClosest}
\alias{monVoyageClosest}
\title{A naive algorithm in the travelling salesman problem}
\usage{
monVoyageClosest(towns)
}
\arguments{
\item{towns}{dataframe of towns generated by the `Towns` function}
}
\value{
vector of town indices of the naive strategy
}
\description{
This algorithms returns the indices of towns with a simple closest-town strategy
}
