% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesman.R
\name{monVoyageR}
\alias{monVoyageR}
\title{The shortest route in the travelling salesman problem}
\usage{
monVoyageR(towns)
}
\arguments{
\item{towns}{dataframe of towns generated by the `Towns` function}
}
\value{
vector of town indices of the shortest path
}
\description{
This algorithms returns the indices of towns in the shortest order for the Euclidean norm
}
