% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monVoyageRcpp}
\alias{monVoyageRcpp}
\title{The shortest route in the travelling salesman problem}
\usage{
monVoyageRcpp(towns)
}
\arguments{
\item{towns}{dataframe of towns generated by the `Towns` function}
}
\description{
This algorithms returns the indices of towns in the shortest order for the Euclidean norm
}
