% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.salesmanRcpp}
\alias{plot.salesmanRcpp}
\title{plot.salesman}
\usage{
\method{plot}{salesmanRcpp}(x, ..., towns, col = "blue")
}
\arguments{
\item{x}{a salesman class object}

\item{...}{Other parameters}

\item{towns}{the dataframe with the towns (X,Y) coordinate}

\item{col}{the color of the path}
}
\value{
plot the towns and the shortest path
}
\description{
Plot the result of the gfpop function with the data-points
}
