% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salesman.R
\name{Towns}
\alias{Towns}
\title{Towns coordinate generator}
\usage{
Towns(n = 7)
}
\arguments{
\item{n}{number of towns to generate}
}
\value{
a dataframe with 2 columns X et Y with the town coordinates
}
\description{
Genearting towns coordinate using a standard normal distribution
}
